requirements.txt
# requirements.txt

# --- 핵심 의존성 ---
streamlit>=1.38.0
google-generativeai>=0.8.0
pandas>=2.2.0
numpy>=1.24.0

# --- MCP (Model Context Protocol) 관련 ---
mcp>=1.13.1
fastmcp>=2.11.0

# --- LangChain 관련 ---
langchain>=0.2.0
langchain-core>=0.2.0
langchain-google-genai>=1.0.0
langchain-mcp-adapters>=0.1.0
langgraph>=0.1.0
langchain-community>=0.2.0 # FAISS, PyPDFLoader 등 RAG 핵심 모듈 포함

# --- RAG 및 벡터스토어 ---
faiss-cpu>=1.8.0      # FAISS 벡터스토어 엔진
pypdf>=4.2.0         # PDF 문서 로더
sentence-transformers 
torch

# --- 이미지 처리 ---
Pillow>=10.0.0

# --- 기타 ---
fastapi>=0.111.0
uvicorn>=0.26.0
matplotlib>=3.7.2